Webhook dispatcher will also lookup the repo
Controller/dispatcher for receiving GitHub webhook will also find the repository and if is not registered in the app (DB), it will end with 404. Otherwise is repo passed to event processors (they would most probably need it) and at the end DB changes are commited, so there is no need to commit in processors.